December/16th-Dec-Tues/previous/requirements.txt
fastapi
sqlalchemy
pymysql
python-decouple